CItest(gRim)	R Documentation

Test for conditional independence

Description

Tests for conditional independence of two variables u and v given a third set W. These tests can be made in a contingency table or in the multivariate normal distribution.

Usage


CItest(x, set, statistic = "dev", method = "chisq", adjust = TRUE,
   slice.info = TRUE, B = 2000, n) 

Arguments

x 	 Either a contingency table (an array or a table) or an empirical covariance matrix. In the latter case the argument n must also be given

set 	 A vector or a formula specifying (u,v,W) – the test being made is for u independent of v given W.

statistic 	 ~~Describe statistic here~~

method 	 ~~Describe method here~~

adjust 	 ~~Describe adjust here~~

slice.info 	 ~~Describe slice.info here~~

B 	 ~~Describe B here~~

n 	 The number of observations (for the empirical covariance matrix)

Details

~~ If necessary, more details than the description above ~~

Value

An object of class ci.test

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

chisq.test

Examples


data(reinis)
CItest(reinis, set=~smoke+systol+phys)


[Package gRim version 0.1.22 Index]