cmod(gRim)	R Documentation

Graphical Gaussian model

Description

Specification of graphical Gaussian model. The 'c' in the name cmod refers to that it is a (graphical) model for 'c'ontinuous variables

Usage


cmod(formula, data = NULL, marginal, fit = TRUE)

Arguments

formula 	 Model specification in one of the following forms: 1) a right-hand sided formula, 2) as a list of generators, 3) an undirected graph (represented either as a graphNEL object or as an adjacency matrix). Notice that there are certain model specification shortcuts, see Section 'details' below

data 	 Data in one of the following forms: 1) A dataframe or 2) a list with elements cov and n.obs (such as returned by the cov.wt() function.)

marginal 	 Should only a subset of the variables be used in connection with the model specification shortcuts

fit 	 Should the model be fitted.

Details

The independence model can be specified as ~.^1 and the saturated model as ~.^.. The marginal argument can be used for specifying the independence or saturated models for only a subset of the variables.

Value

An object of class cmod (a list)

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

dmod

Examples


## Graphical Gaussian model
data(carcass)
cm1<-cmod(~.^., carcass)
cm2<-backward(cm1,k=log(nrow(carcass)),
   fixin=list(c("LMP","M11","M12","M13"),c("LMP","F11","F12","F13"))) 
cm3<-backward(cm1,k=log(nrow(carcass)))

[Package gRim version 0.1.22 Index]