dmod(gRim)	R Documentation

Log–linear model

Description

Specification of log–linear (graphical) model. The 'd' in the name dmod refers to that it is a (graphical) model for 'd'iscrete variables

Usage


dmod(formula, data = NULL, marginal, context,
 interactions=NULL, fit = TRUE)

Arguments

formula 	 Model specification in one of the following forms: 1) a right-hand sided formula, 2) as a list of generators, 3) an undirected graph (represented either as a graphNEL object or as an adjacency matrix). Notice that there are certain model specification shortcuts, see Section 'details' below.

data 	 A table

context 	 Restrict attention to a (multidimensional) slice of the table; the slide being given by specific levels of certain variables. See 'examples' below.

interactions 	 A number given the highest order interactions in the model, see Section 'details' below.

marginal 	 Should only a subset of the variables be used in connection with the model specification shortcuts

fit 	 Should the model be fitted.

Details

The independence model can be specified as ~.^1 and the saturated model as ~.^.. The marginal argument can be used for specifying the independence or saturated models for only a subset of the variables.

Setting e.g. interactions=3 implies that there will be at most three factor interactions in the model.

Value

An object of class dmod (a list)

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

cmod

Examples



## Graphical log-linear model
data(reinis)
dm1<-dmod(~.^., reinis)
dm2<-backward(dm1, k=2)
dm3<-backward(dm1, k=2, fixin=list(c("family","phys","systol")))
## At most 3-factor interactions
dm1<-dmod(~.^., data=reinis,interactions=3)

## Define the independence model for only a subset of the variables
## in a specific context. 
dm1 <- dmod(~.^., margin=c("smoke","mental","phys","systol"),
data=reinis, context=list("family", "y"))


[Package gRim version 0.1.22 Index]