evalInEdges; evalOutEdges(gRim)	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


evalInEdges (object, edgeList=findInEdges(object), criterion = "aic", k = 2,
 alpha = if (criterion=="aic") 0 else 0.05, headlong = FALSE, details = 1, ...)
evalOutEdges(object, edgeList=findOutEdges(object), criterion = "aic", k = 2,
 alpha = if (criterion=="aic") 0 else 0.05, headlong = FALSE, details = 1, ...)

Arguments

object 	 ~~Describe object here~~

edgeList 	 ~~Describe edgeList here~~

criterion 	 ~~Describe criterion here~~

k 	 ~~Describe k here~~

alpha 	 ~~Describe alpha here~~

headlong 	 ~~Describe headlong here~~

details 	 ~~Describe print here~~

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

findInEdges,findOutEdges, testdelete,testdelete,

Examples


data(math)
cm1 <- cmod(~me:ve+ve:al+al:an, data=math)
evalInEdges(cm1)
evalOutEdges(cm1)

[Package gRim version 0.1.22 Index]