findInEdges,findOutEdges(gRim)	R Documentation

Find edges in (or not in) a graph

Description

Find edges in (or not in) a graph. It is possible to restrict the set of edges to be returned.

Usage


findInEdges(x, type = "unrestricted", ...)
findOutEdges(x, type = "unrestricted", ...)

Arguments

x 	 Either 1) An imod model object, 2) a list containing the generating class for a graph or 3) a graphNEL graph

type 	 Either "unrestricted" or "decomposable"; see Section 'details' below.

... 	 Further arguments; currently not used.

Details

findInEdges(): If type is "decomposable", then the edges returned are as follows: An edge e in the graph is returned iff deleting e from the graph given by the object x gives a new graph which is decomposable.

findOutEdges(): If type is "decomposable", then the edges returned are as follows: An edge e not in the graph is returned iff adding e to the graph given by the object x gives a new graph which is decomposable.

Value

A list of edges.

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

evalInEdges, evalOutEdges

Examples


## findInEdges
##
gg1 <- ug(~a:b:c+a:c:d)
## plot(gg1)
findInEdges(gg1)
findInEdges(gg1, type="decomposable")

## findOutEdges
##
gg2 <- ug(~a:b+b:c+c:d)
## plot(gg2)
findOutEdges(gg2)
findOutEdges(gg2, type="decomposable")

[Package gRim version 0.1.22 Index]