ggmfit(gRim)	R Documentation

Iterative proportional fitting of graphical Gaussian model

Description

Fit graphical Gaussian model by iterative proportional fitting.

Usage


ggmfit(S, nobs, glist, start = NULL, eps = 0.001, iter = 20,
 print = 1,...)
ggmfitr(S, nobs, glist, start = NULL, eps = 0.001, iter = 20,
 print = 1,...)

Arguments

S 	 Empirical covariance matrix

nobs 	 Number of observations

glist 	 Generating class for model (a list)

start 	 Initial value for concentration matrix

eps 	 Convergence criterion

iter 	 Maximum number of iterations

print 	 If TRUE, details from the estimation are printed

... 	 Optional arguments; currently not used

Details

ggmfit is based on a C implementation. ggmfitr is implemented purely in R (and is provided mainly as a benchmark for the C-version).

Value

A list with

lrt 	 Likelihood ratio statistic (-2logL)

df 	 Degrees of freedom

logL 	 log likelihood

K 	 Estimated concentration matrix (inverse covariance matrix)

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

References

FIXME

See Also

loglin

Examples


## FIXME: Output fra de to funktioner passer ikke helt med hinanden
data(math)
d <- cov.wt(math, method="ML")
glist <- list(c("al","st","an"), c("me","ve","al"))
ggmfit(d$cov, d$n.obs, glist)
ggmfitr(d$cov, d$n.obs, glist)

[Package gRim version 0.1.27 Index]