loglinDim(gRim)	R Documentation

Return the dimension of a log-linear model

Description

Return the dimension of a log-linear model given by the generating class 'glist'. If the model is decomposable and adjusted dimension can be found.

Usage


loglinDim(glistNUM, table)
loglinDecDim(glistNUM, table, adjust=TRUE)

Arguments

glistNUM 	 ~~Describe glistNUM here~~

table 	 ~~Describe dtab here~~

adjust

Details

If the model is decomposable and adjusted dimension can be found (using loglinDecDim()).

Value

A numeric.

Author(s)

Søren Højsgaard, sorenh@agrsci.dk

See Also

dmod, glm, loglm

Examples


## To be done
## Create example with glm and check dimension of model matrix.

[Package gRim version 0.1.27 Index]