stepwise.imod(gRim)	R Documentation

Stepwise model selection in (graphical) interaction models

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage


## S3 method for class 'imod':
stepwise(object, direction = "backward",
 criterion = "aic", alpha = if (criterion == "aic") 0 else 0.05,
 type ="decomposable", headlong = FALSE, steps = 1000, k = 2,
 fixin=NULL, fixout=NULL,details = 0, trace = 2, ...)
backward(object, criterion = "aic",
 alpha = if (criterion == "aic") 0 else 0.05, type = "decomposable",
 headlong = FALSE, steps = 1000, k = 2, fixin=NULL, details = 1, trace = 2,...)
forward(object,  criterion = "aic",
 alpha = if (criterion == "aic") 0 else 0.05, type = "decomposable",
 headlong = FALSE, steps = 1000, k = 2, fixout=NULL,  details = 1, trace = 2,...)

Arguments

object 	 ~~Describe object here~~

direction 	 ~~Describe direction here~~

type 	 ~~Describe type here~~

criterion 	 ~~Describe criterion here~~

alpha 	 ~~Describe alpha here~~

headlong 	 ~~Describe headlong here~~

steps 	 ~~Describe steps here~~

k 	 ~~Describe k here~~

details 	 ~~Describe print here~~

trace 	 ~~Describe trace here~~

fixin

fixout

... 	 ~~Describe ... here~~

Details

~~ If necessary, more details than the description above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of 'comp1'

comp2 	 Description of 'comp2'

...

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples


  data(reinis)
## The saturated model
m1 <- dmod(~.^., data=reinis)
m2 <- stepwise(m1)
m2


[Package gRim version 0.1.23 Index]